Bug fix in RNF model
Force long type for inputs - if one of the examples is empty and happens
to be first in the batch, pad_sequence() infers the type as floatTensor
instead of LongTensor - ensure all inputs are of type long to avoid
error in embedding layer

+++ text_classification/encoders/encoders.py
@@ -25,7 +25,7 @@ class RNFEncoder(BaseEncoder):
         # pad the sequences
         x = pad_sequence(data, batch_first=True)
 
-        return x, torch.Tensor(targets).long()
+        return x.long(), torch.Tensor(targets).long()
 
     def _encode(self, example):
